Adding html task to gulpfile.
This task moves html files from the src folder to the public_html folder. This
means there are two copies of the same file, but it makes more sense to develop
all the files in one place (src folder), and then have them all in their final
form in the public_html folder.