Adjusting the way in which hero images account for their height.
Previously, in order to compensate for the full page height of a hero image,
the 'main' element was given a padding of 100vh. This was bad because it meant
that all pages had a padding of 100vh at the beginning of the 'main' element.
For now, hero images are wrapped with a div that has the 'hero-wrapper' class,
which has a height of 100vh and affects other elements.